Sorted vector special case: default first element
A lot of code using vectors in LO relies on the fact, that the
first entry in the vector contains the default value.

Therefore this adds a boolean to the constructor, which leaves
the first entry unsorted in the vector and special cases find
and insert.

All other functions are unaffected.

Change-Id: I9603f47be4fb56d991f42066ce9f5ad0ab6ffdf8