Change SwPageDescs to o3tl::sorted_vector like API
This moves the SwPageDescs class to pagedesc.hxx, changes the
std::vector inheritance to private and extends the class to a
o3tl::sorted_vector compatible API.
The insert internally uses still uses push_back.

This also drops multiple implementations for the "find SwPageDesc
by name algorithm".

This should just be a cleanup patch and is a preparation for the
change of SwPageDescs from vector to o3tl::sorted_vector.

Change-Id: I08561c139d1af9bad6ca68be27466fac33fab384