opencollada: uninitialized bool variable broke the parsing
GeneratedSaxParser::Utils:toURI() method has a bool output
parameter called failed, which is assumed to be set inside
the method before return. At some place the caller code does
not initialize the bool variable passed to this failed parameter
and so when the caller checkes the returned value it is
undefined.
e.g. it can be false when the method called successfully so
the error handling throws away the returned URI.

Change-Id: I3f90fab657a86b42bba0f492518e36c343e69d21